Settings: Move custom vib pattern to its own preference
Also improve the code while at it
