+++ res/xml/custom_vibration_pattern.xml
@@ -1,0 +1,50 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- Copyright (C) 2020 Yet Another AOSP Project
+
+     Licensed under the Apache License, Version 2.0 (the "License");
+     you may not use this file except in compliance with the License.
+     You may obtain a copy of the License at
+
+          http://www.apache.org/licenses/LICENSE-2.0
+
+     Unless required by applicable law or agreed to in writing, software
+     distributed under the License is distributed on an "AS IS" BASIS,
+     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+     See the License for the specific language governing permissions and
+     limitations under the License.
+-->
+
+<PreferenceScreen
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:settings="http://schemas.android.com/apk/res-auto"
+    android:title="@string/vibration_pattern_title"
+    android:key="custom_vibration_pattern">
+
+    <!-- Custom ringtone vibration pattern 1 -->
+    <com.crdroid.settings.preferences.CustomSeekBarPreference
+        android:key="custom_vibration_pattern1"
+        android:order="-123"
+        android:max="1000"
+        settings:min="0"
+        settings:interval="100"
+        settings:units="ms"/>
+
+    <!-- Custom ringtone vibration pattern 2 -->
+    <com.crdroid.settings.preferences.CustomSeekBarPreference
+        android:key="custom_vibration_pattern2"
+        android:order="-122"
+        android:max="1000"
+        settings:min="0"
+        settings:interval="100"
+        settings:units="ms"/>
+
+    <!-- Custom ringtone vibration pattern 3 -->
+    <com.crdroid.settings.preferences.CustomSeekBarPreference
+        android:key="custom_vibration_pattern3"
+        android:order="-121"
+        android:max="1000"
+        settings:min="0"
+        settings:interval="100"
+        settings:units="ms"/>
+
+</PreferenceScreen>
