+++ res/xml/sound_settings.xml
@@ -173,47 +173,16 @@
         android:title="@string/vibration_pattern_title"
         android:order="-85"
         android:entries="@array/vibration_pattern_entries"
-        android:entryValues="@array/vibration_pattern_values"
-        android:dependency="vibrate_for_calls"/>
+        android:entryValues="@array/vibration_pattern_values"/>
 
-    <!-- Custom ringtone vibration pattern category -->
-    <PreferenceCategory
+    <!-- Custom ringtone vibration pattern preference -->
+    <Preference
         android:key="custom_vibration_pattern"
         android:title="@string/custom_vibration_pattern_title"
-        android:order="-84"
-        android:visibility="gone">
-
-        <!-- Custom ringtone vibration pattern 1 -->
-        <com.crdroid.settings.preferences.CustomSeekBarPreference
-            android:key="custom_vibration_pattern1"
-            android:order="-83"
-            android:max="1000"
-            settings:min="0"
-            settings:interval="100"
-            settings:units="ms"
-            android:dependency="vibrate_for_calls"/>
-
-        <!-- Custom ringtone vibration pattern 2 -->
-        <com.crdroid.settings.preferences.CustomSeekBarPreference
-            android:key="custom_vibration_pattern2"
-            android:order="-82"
-            android:max="1000"
-            settings:min="0"
-            settings:interval="100"
-            settings:units="ms"
-            android:dependency="vibrate_for_calls"/>
-
-        <!-- Custom ringtone vibration pattern 3 -->
-        <com.crdroid.settings.preferences.CustomSeekBarPreference
-            android:key="custom_vibration_pattern3"
-            android:order="-81"
-            android:max="1000"
-            settings:min="0"
-            settings:interval="100"
-            settings:units="ms"
-            android:dependency="vibrate_for_calls"/>
-
-    </PreferenceCategory>
+        android:summary="@string/custom_vibration_pattern_summary"
+        android:order="-83"
+        android:fragment="com.android.settings.sound.CustomVibrationPreferenceFragment"
+        android:visibility="gone"/>
 
     <com.android.settingslib.PrimarySwitchPreference
         android:key="gesture_prevent_ringing_sound"
