+++ src/com/android/settings/notification/VibrationPatternPreferenceController.java
@@ -30,9 +30,8 @@ import androidx.preference.PreferenceCategory;
 import androidx.preference.PreferenceScreen;
 
 import com.android.settings.R;
+import com.android.settings.Utils;
 import com.android.settingslib.core.AbstractPreferenceController;
-
-import com.crdroid.settings.preferences.CustomSeekBarPreference;
 
 /**
  * This class allows choosing a vibration pattern while ringing
@@ -42,15 +41,9 @@ public class VibrationPatternPreferenceController extends AbstractPreferenceCont
 
     private static final String KEY_VIB_PATTERN = "vibration_pattern";
     private static final String KEY_CUSTOM_VIB_CATEGORY = "custom_vibration_pattern";
-    private static final String KEY_CUSTOM_VIB1 = "custom_vibration_pattern1";
-    private static final String KEY_CUSTOM_VIB2 = "custom_vibration_pattern2";
-    private static final String KEY_CUSTOM_VIB3 = "custom_vibration_pattern3";
 
     private ListPreference mVibPattern;
-    private PreferenceCategory mCustomVibCategory;
-    private CustomSeekBarPreference mCustomVib1;
-    private CustomSeekBarPreference mCustomVib2;
-    private CustomSeekBarPreference mCustomVib3;
+    private Preference mCustomVibCategory;
 
     private static class VibrationEffectProxy {
         public VibrationEffect createWaveform(long[] timings, int[] amplitudes, int repeat) {
@@ -133,7 +126,7 @@ public class VibrationPatternPreferenceController extends AbstractPreferenceCont
 
     @Override
     public boolean isAvailable() {
-        return true;
+        return Utils.isVoiceCapable(mContext);
     }
 
     @Override
@@ -151,11 +144,8 @@ public class VibrationPatternPreferenceController extends AbstractPreferenceCont
         mVibPattern.setSummary(mVibPattern.getEntries()[vibPattern]);
         mVibPattern.setOnPreferenceChangeListener(this);
 
-        mCustomVibCategory = (PreferenceCategory) screen.findPreference(KEY_CUSTOM_VIB_CATEGORY);
-        mCustomVib1 = (CustomSeekBarPreference) screen.findPreference(KEY_CUSTOM_VIB1);
-        mCustomVib2 = (CustomSeekBarPreference) screen.findPreference(KEY_CUSTOM_VIB2);
-        mCustomVib3 = (CustomSeekBarPreference) screen.findPreference(KEY_CUSTOM_VIB3);
-        updateCustomVibVisibility(vibPattern == 5);
+        mCustomVibCategory = screen.findPreference(KEY_CUSTOM_VIB_CATEGORY);
+        mCustomVibCategory.setVisible(vibPattern == 5);
     }
 
     @Override
@@ -166,58 +156,11 @@ public class VibrationPatternPreferenceController extends AbstractPreferenceCont
                     Settings.System.RINGTONE_VIBRATION_PATTERN, vibPattern);
             mVibPattern.setSummary(mVibPattern.getEntries()[vibPattern]);
             boolean isCustom = vibPattern == 5;
-            updateCustomVibVisibility(isCustom);
+            mCustomVibCategory.setVisible(isCustom);
             if (!isCustom) previewPattern();
-            return true;
-        } else if (preference == mCustomVib1) {
-            updateCustomVib(0, (Integer) newValue);
-            return true;
-        } else if (preference == mCustomVib2) {
-            updateCustomVib(1, (Integer) newValue);
-            return true;
-        } else if (preference == mCustomVib3) {
-            updateCustomVib(2, (Integer) newValue);
             return true;
         }
         return false;
-    }
-
-    private void updateCustomVibVisibility(boolean show) {
-        mCustomVibCategory.setVisible(show);
-        mCustomVib1.setVisible(show);
-        mCustomVib2.setVisible(show);
-        mCustomVib3.setVisible(show);
-        if (show) updateCustomVibPreferences();
-    }
-
-    private void updateCustomVibPreferences() {
-        String value = Settings.System.getString(mContext.getContentResolver(),
-                Settings.System.CUSTOM_RINGTONE_VIBRATION_PATTERN);
-        if (value != null) {
-            String[] customPattern = value.split(",", 3);
-            mCustomVib1.setValue(Integer.parseInt(customPattern[0]));
-            mCustomVib2.setValue(Integer.parseInt(customPattern[1]));
-            mCustomVib3.setValue(Integer.parseInt(customPattern[2]));
-        } else { // set default
-            mCustomVib1.setValue(0);
-            mCustomVib2.setValue(800);
-            mCustomVib3.setValue(800);
-            Settings.System.putString(mContext.getContentResolver(),
-                    Settings.System.CUSTOM_RINGTONE_VIBRATION_PATTERN, "0,800,800");
-        }
-        mCustomVib1.setOnPreferenceChangeListener(this);
-        mCustomVib2.setOnPreferenceChangeListener(this);
-        mCustomVib3.setOnPreferenceChangeListener(this);
-    }
-
-    private void updateCustomVib(int index, int value) {
-        String[] customPattern = Settings.System.getString(mContext.getContentResolver(),
-                Settings.System.CUSTOM_RINGTONE_VIBRATION_PATTERN).split(",", 3);
-        customPattern[index] = String.valueOf(value);
-        Settings.System.putString(mContext.getContentResolver(),
-                Settings.System.CUSTOM_RINGTONE_VIBRATION_PATTERN, String.join(
-                ",", customPattern[0], customPattern[1], customPattern[2]));
-        previewPattern();
     }
 
     private void previewPattern() {
@@ -241,22 +184,6 @@ public class VibrationPatternPreferenceController extends AbstractPreferenceCont
                 break;
             case 4:
                 effect = vibrationEffectProxy.createWaveform(DA_DZZZ_DA_VIBRATION_PATTERN,
-                        SEVEN_ELEMENTS_VIBRATION_AMPLITUDE, -1);
-                break;
-            case 5:
-                String[] customVib = Settings.System.getString(
-                        mContext.getContentResolver(),
-                        Settings.System.CUSTOM_RINGTONE_VIBRATION_PATTERN).split(",", 3);
-                long[] customVibPattern = {
-                    0, // No delay before starting
-                    Long.parseLong(customVib[0]), // How long to vibrate
-                    400, // Delay
-                    Long.parseLong(customVib[1]), // How long to vibrate
-                    400, // Delay
-                    Long.parseLong(customVib[2]), // How long to vibrate
-                    400, // How long to wait before vibrating again
-                };
-                effect = vibrationEffectProxy.createWaveform(customVibPattern,
                         SEVEN_ELEMENTS_VIBRATION_AMPLITUDE, -1);
                 break;
             default:
